requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-autoslug==1.9.8
django-countries==7.5.1
django-environ==0.9.0
django-filter==22.1
django-model-utils==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
progressbar2==4.2.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-utils==3.5.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
swapper==1.3.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.14
